Update keystone script

+++ openstack_keystone.sh
@@ -29,9 +29,11 @@ EOF
 . ./stackrc
 
 # edit keystone conf file to use templates and mysql
-cp /etc/keystone/keystone.conf /etc/keystone/keystone.conf.orig
+if [ ! -f /etc/keystone/keystone.conf.orig ]; then
+	cp /etc/keystone/keystone.conf /etc/keystone/keystone.conf.orig
+fi
 sed -e "
-/^admin_token = ADMIN/s/^.*$/admin_token = $token/
+/^#\?\s*admin_token = ADMIN/s/^.*$/admin_token = $token/
 /^driver = keystone.catalog.backends.sql.Catalog/d
 /^\[catalog\]/a driver = keystone.catalog.backends.templated.TemplatedCatalog 
 /^\[catalog\]/a template_file = /etc/keystone/default_catalog.templates
@@ -82,19 +84,19 @@ ANOTHER_ROLE=$(get_id keystone role-create --name=anotherrole)
 
 
 # Add Roles to Users in Tenants
-keystone user-role-add --user $ADMIN_USER --role $ADMIN_ROLE --tenant_id $ADMIN_TENANT
-keystone user-role-add --user $ADMIN_USER --role $ADMIN_ROLE --tenant_id $DEMO_TENANT
-keystone user-role-add --user $DEMO_USER --role $ANOTHER_ROLE --tenant_id $DEMO_TENANT
+keystone user-role-add --user_id $ADMIN_USER --role_id $ADMIN_ROLE --tenant_id $ADMIN_TENANT
+keystone user-role-add --user_id $ADMIN_USER --role_id $ADMIN_ROLE --tenant_id $DEMO_TENANT
+keystone user-role-add --user_id $DEMO_USER --role_id $ANOTHER_ROLE --tenant_id $DEMO_TENANT
 
 # TODO(termie): these two might be dubious
-keystone user-role-add --user $ADMIN_USER --role $KEYSTONEADMIN_ROLE --tenant_id $ADMIN_TENANT
-keystone user-role-add --user $ADMIN_USER --role $KEYSTONESERVICE_ROLE --tenant_id $ADMIN_TENANT
+keystone user-role-add --user_id $ADMIN_USER --role_id $KEYSTONEADMIN_ROLE --tenant_id $ADMIN_TENANT
+keystone user-role-add --user_id $ADMIN_USER --role_id $KEYSTONESERVICE_ROLE --tenant_id $ADMIN_TENANT
 
 
 # The Member role is used by Horizon and Swift so we need to keep it:
 MEMBER_ROLE=$(get_id keystone role-create --name=Member)
-keystone user-role-add --user $DEMO_USER --role $MEMBER_ROLE --tenant_id $DEMO_TENANT
-keystone user-role-add --user $DEMO_USER --role $MEMBER_ROLE --tenant_id $INVIS_TENANT
+keystone user-role-add --user_id $DEMO_USER --role_id $MEMBER_ROLE --tenant_id $DEMO_TENANT
+keystone user-role-add --user_id $DEMO_USER --role_id $MEMBER_ROLE --tenant_id $INVIS_TENANT
 
 
 # Configure service users/roles
@@ -103,16 +105,16 @@ NOVA_USER=$(get_id keystone user-create --name=nova \
                                         --tenant_id $SERVICE_TENANT \
                                         --email=$email)
 keystone user-role-add --tenant_id $SERVICE_TENANT \
-                       --user $NOVA_USER \
-                       --role $ADMIN_ROLE
+                       --user_id $NOVA_USER \
+                       --role_id $ADMIN_ROLE
 
 GLANCE_USER=$(get_id keystone user-create --name=glance \
                                           --pass="$SERVICE_PASSWORD" \
                                           --tenant_id $SERVICE_TENANT \
                                           --email=$email)
 keystone user-role-add --tenant_id $SERVICE_TENANT \
-                       --user $GLANCE_USER \
-                       --role $ADMIN_ROLE
+                       --user_id $GLANCE_USER \
+                       --role_id $ADMIN_ROLE
 
 if [[ "$ENABLED_SERVICES" =~ "swift" ]]; then
     SWIFT_USER=$(get_id keystone user-create --name=swift \
@@ -120,8 +122,8 @@ if [[ "$ENABLED_SERVICES" =~ "swift" ]]; then
                                              --tenant_id $SERVICE_TENANT \
                                              --email=$email)
     keystone user-role-add --tenant_id $SERVICE_TENANT \
-                           --user $SWIFT_USER \
-                           --role $ADMIN_ROLE
+                           --user_id $SWIFT_USER \
+                           --role_id $ADMIN_ROLE
     # Nova needs ResellerAdmin role to download images when accessing
     # swift through the s3 api. The admin role in swift allows a user
     # to act as an admin for their tenant, but ResellerAdmin is needed
@@ -129,8 +131,8 @@ if [[ "$ENABLED_SERVICES" =~ "swift" ]]; then
     # configurable in swift-proxy.conf
     RESELLER_ROLE=$(get_id keystone role-create --name=ResellerAdmin)
     keystone user-role-add --tenant_id $SERVICE_TENANT \
-                           --user $NOVA_USER \
-                           --role $RESELLER_ROLE
+                           --user_id $NOVA_USER \
+                           --role_id $RESELLER_ROLE
 fi
 
 if [[ "$ENABLED_SERVICES" =~ "quantum" ]]; then
@@ -139,8 +141,8 @@ if [[ "$ENABLED_SERVICES" =~ "quantum" ]]; then
                                                --tenant_id $SERVICE_TENANT \
                                                --email=$email)
     keystone user-role-add --tenant_id $SERVICE_TENANT \
-                           --user $QUANTUM_USER \
-                           --role $ADMIN_ROLE
+                           --user_id $QUANTUM_USER \
+                           --role_id $ADMIN_ROLE
 fi
 
 echo "######################################################################################"
